requirements.txt
langchain
transformers
pinecone-client
python-dotenv
pyyaml
sentence-transformers
flask 